fix: wire settings to localStorage, fix ChirpStack ingest pipeline
- Settings page: remove dead settings, wire defaultRange/autoRefresh/units/timezone to localStorage via new frontend/lib/settings.ts
- Dashboard + GaugeCard: read units/timezone/autoRefresh from settings, pass to charts
- DetailChart + gauge detail page: metric conversion, timezone-aware labels, settings-driven refresh
- alert_checker: open own DB session instead of reusing closed request session
- models: widen gauge.device_id from SmallInteger to Integer (real device uses 17039361)
- ingest: remove db arg from create_task call
- next.config.mjs: add /health rewrite for backend status check
- Add Chirpstack_dashboard.png and single-log9.json to media

+++ backend/app/models.py
@@ -10,7 +10,7 @@ class Gauge(Base):
     __tablename__ = "gauges"
 
     id = Column(Integer, primary_key=True, index=True)
-    device_id = Column(SmallInteger, unique=True, nullable=False)
+    device_id = Column(Integer, unique=True, nullable=False)
     name = Column(String(255), nullable=False, default="Unknown Device")
     town_state = Column(String(255))
     last_updated_at = Column(DateTime(timezone=True))
